+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 ## [0.29.0] - 2025-05-19
 ### Added
 - View recurring transaction details with edit and delete actions.
+- View upcoming recurring transactions in a new screen.
 
 ## [0.28.0] - 2025-05-19
 ### Added